@@ -1,56 +1,16 @@
 <?php
 
-global $project;
-$project = 'mysite';
+use SilverStripe\Security\PasswordValidator;
+use SilverStripe\Security\Member;
+use SilverStripe\Control\Director;
+// remove PasswordValidator for SilverStripe 5.0
+$validator = new PasswordValidator();
 
-global $database;
-//You have to set up your db.
-$database = 'riverrun';
- 
-// Use _ss_environment.php file for configuration
-require_once("conf/ConfigureFromEnv.php");
-SSViewer::set_theme('bijou'); 
+$validator->minLength(8);
+$validator->checkHistoricalPasswords(6);
+Member::set_password_validator($validator);
 
-// Set the site locale
-i18n::set_locale('en_US');
-FulltextSearchable::enable();
-// Enable nested URLs for this site (e.g. page/sub-page/)
-if (class_exists('SiteTree')) SiteTree::enable_nested_urls();
-
-Requirements::set_force_js_to_bottom(true);
-
-// add a button to remove formatting
-HtmlEditorConfig::get('cms')->insertButtonsBefore(
-    'styleselect',
-    'removeformat'
-);
-
-// tell the button which tags it may remove
-HtmlEditorConfig::get('cms')->setOption(
-    'removeformat_selector',
-    'b,strong,em,i,span,ins'
-);
-
-//remove font->span conversion
-
-HtmlEditorConfig::get('cms')->setOption(
-	'convert_fonts_to_spans', 'false,'
-);
-
-HtmlEditorConfig::get('cms')->setOptions(array(
-'valid_elements' => "@[id|class|style|title],#a[id|rel|rev|dir|tabindex|accesskey|type|name|href|target|title|class],-strong/-b[class],-em/-i[class],-strike[class],-u[class],#p[id|dir|class|align|style],-ol[class],-ul[class],-li[class],br,img[id|dir|longdesc|usemap|class|src|border|alt=|title|width|height|align],-sub[class],-sup[class],-blockquote[dir|class],-table[border=0|cellspacing|cellpadding|width|height|class|align|summary|dir|id|style],-tr[id|dir|class|rowspan|width|height|align|valign|bgcolor|background|bordercolor|style],tbody[id|class|style],thead[id|class|style],tfoot[id|class|style],#td[id|dir|class|colspan|rowspan|width|height|align|valign|scope|style],-th[id|dir|class|colspan|rowspan|width|height|align|valign|scope|style],caption[id|dir|class],-h1[id|dir|class|align|style],-h2[id|dir|class|align|style],-h3[id|dir|class|align|style],-h4[id|dir|class|align|style],-h5[id|dir|class|align|style],-h6[id|dir|class|align|style],hr[class],dd[id|class|title|dir],dl[id|class|title|dir],dt[id|class|title|dir],@[id,style,class],small",
-'extended_valid_elements' => "img[class|src|alt|title|hspace|vspace|width|height|align|onmouseover|onmouseout|name|usemap],#iframe[src|name|width|height|align|frameborder|marginwidth|marginheight|scrolling],object[width|height|data|type],param[name|value],map[class|name|id],area[shape|coords|href|target|alt]"
-));
-
-// TinyMCE cleanup on paste
-HtmlEditorConfig::get('cms')->setOption('paste_auto_cleanup_on_paste','true');
-HtmlEditorConfig::get('cms')->setOption('paste_remove_styles','true');
-HtmlEditorConfig::get('cms')->setOption('paste_remove_styles_if_webkit','true');
-HtmlEditorConfig::get('cms')->setOption('paste_strip_class_attributes','true');
-GD::set_default_quality(80);
 
 if(Director::isLive()) {
 	Director::forceSSL();
 }
-Authenticator::unregister('MemberAuthenticator');
-Authenticator::set_default_authenticator('SAMLAuthenticator');
